9.1.2 Made new directory api abd move to there api.js.
